anotation-targets: Add support for SVG selectors.
All SVG selectors are parsed (if a DOM parser is available) and, in
addition to their SVG source, transformed into a set of points, as well
as a rectangle encompassing the selector. This is to help clients
work with the selectors, since points and rectangles are easier to
handle and calculate with than the full SVG.

One additional thing that should probably be done is to
massage/normalize the actual SVG source code to reduce the number of
variations that clients will have to handle.